% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAExpSuv.R
\docType{package}
\name{IAExpSuv}
\alias{IAExpSuv}
\title{IAExpSuv}
\description{
IAExpSuv is used for for predicting the five-year risk of death for a single or a group of stage IA lung adenocarcinoma patients based on gene expression data
}
