% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAPrem.R
\name{IAprem}
\alias{IAprem}
\title{IAprem}
\usage{
IAprem(x)
}
\arguments{
\item{x}{a data.frame containing GeneID and gene expression of a group of patients}
}
\description{
IAprem is used for predicting the five-year risk of death for a group of stage IA lung adenocarcinoma patients
}
