% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IAPres.R
\name{IApres}
\alias{IApres}
\title{IApres}
\usage{
IApres(x)
}
\arguments{
\item{x}{a data.frame containing GeneID and gene expression of a single patient}
}
\description{
IApres is used for predicting the five-year risk of death for a single stage IA lung adenocarcinoma patient
}
